pyproject.toml
[project]
name = "relay"
version = "0.1.0"
description = "Simple yet effective open source LLM Studio."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "greenlet>=3.2.4",
    "loguru>=0.7.3",
    "redis[hiredis]>=6.4.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.20",
    "pillow>=11.3.0",
    "mcp[cli]>=1.14.0",
    "llm-registry>=0.4.2",
    "alembic-postgresql-enum>=1.8.0",
    "logfire[fastapi]>=4.7.0",
    "pydantic-ai>=1.0.6",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pre-commit>=4.3.0",
    "ruff>=0.13.0",
    "tomli-w>=1.2.0",
]

[tool.codespell]
skip = ".git*,*.svg,*.lock,*.css,*.yaml"
check-hidden = true
ignore-words-list = ['selectin']
